Updated return types on SQL functions. Fixes #29.

+++ packages/server/src/db/migrations/up_000.sql
@@ -134,7 +134,7 @@ FOR EACH ROW
 EXECUTE PROCEDURE history_update();
 
 CREATE OR REPLACE FUNCTION get_last_week_experience(username TEXT)
-RETURNS TABLE(experience BIGINT)
+RETURNS TABLE(experience INT)
 AS $BODY$
     SELECT SUM (h.experience)
     FROM history h
@@ -145,7 +145,7 @@ $BODY$
 LANGUAGE 'sql';
 
 CREATE OR REPLACE FUNCTION get_last_month_experience(username TEXT)
-RETURNS TABLE(experience BIGINT)
+RETURNS TABLE(experience INT)
 AS $BODY$
     SELECT SUM (h.experience)
     FROM history h
@@ -156,7 +156,7 @@ $BODY$
 LANGUAGE 'sql';
 
 CREATE OR REPLACE FUNCTION get_current_month_experience(username TEXT)
-RETURNS TABLE(experience BIGINT)
+RETURNS TABLE(experience INT)
 AS $BODY$
     SELECT SUM (h.experience)
     FROM history h
@@ -166,7 +166,7 @@ $BODY$
 LANGUAGE 'sql';
 
 CREATE OR REPLACE FUNCTION get_current_week_experience(username TEXT)
-RETURNS TABLE(experience BIGINT)
+RETURNS TABLE(experience INT)
 AS $BODY$
     SELECT SUM (h.experience)
     FROM history h
